[test/plugin]support plugin testcases run in docker-compose (#3612)
* support docker-compose

* sync

* support docker-compose

* force rm container

* solrj-7.x-scenario

* fix license header

* fix actualData.yml

* verify Jenkins run on distribution

* test Jenkinsfile v2

* test Jenkinsfile v3, God bless

* test V4, God Bless

* test v5

* rollback

* update opertaion of post
