+++ Jenkinsfile-Agent-Test-2
@@ -58,36 +58,18 @@ pipeline {
             parallel {
                 stage('Group1') {
                     stages {
-                        stage('httpclient 4.3.x-4.5.x (14)') {
+                        stage('dubbo 2.5.x-2.6.x (10)') {
                             steps {
-                                sh 'bash test/plugin/run.sh --build_id=${BUILD_ID} httpclient-4.3.x-scenario'
-                            }
-                        }
-
-                        stage('ehcache 2.8.x-2.10.x (19)') {
-                            steps {
-                                sh 'bash test/plugin/run.sh --build_id=${BUILD_ID} ehcache-2.x-scenario'
-                            }
-                        }
-
-                        stage('dubbo 2.7.x (4)') {
-                            steps {
-                                sh 'bash test/plugin/run.sh --build_id=${BUILD_ID} dubbo-2.7.x-scenario'
-                            }
-                        }
-
-                        stage('spring async 4.3.x-5.1.x (35)'){
-                            steps{
-                                sh 'bash test/plugin/run.sh --build_id=${BUILD_ID} spring-async-scenario'
+                                sh 'bash test/plugin/run.sh --build_id=${BUILD_ID} dubbo-2.5.x-scenario'
                             }
                         }
                     }
                 }
                 stage('Group2') {
                     stages {
-                        stage('jetty 9.x (63)') {
+                        stage('solrj 7.x (12)') {
                             steps {
-                                sh 'bash test/plugin/run.sh --build_id=${BUILD_ID} jetty-scenario'
+                                sh 'bash test/plugin/run.sh --build_id=${BUILD_ID} solrj-7.x-scenario'
                             }
                         }
                     }
