+++ test/plugin/runner-helper/src/main/java/org/apache/skywalking/plugin/test/helper/vo/CaseConfiguration.java
@@ -1,0 +1,113 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.skywalking.plugin.test.helper.vo;
+
+import java.util.List;
+import java.util.Map;
+
+public class CaseConfiguration {
+    private String type;
+    private String entryService;
+    private String healthCheck;
+    private String framework;
+    private String hostname;
+    private String startScript;
+    private Map<String, DependencyComponent> dependencies;
+    private List<String> environment;
+    private List<String> expose;
+    private List<String> depends_on;
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+
+    public String getEntryService() {
+        return entryService;
+    }
+
+    public void setEntryService(String entryService) {
+        this.entryService = entryService;
+    }
+
+    public String getHealthCheck() {
+        return healthCheck;
+    }
+
+    public void setHealthCheck(String healthCheck) {
+        this.healthCheck = healthCheck;
+    }
+
+    public String getFramework() {
+        return framework;
+    }
+
+    public void setFramework(String framework) {
+        this.framework = framework;
+    }
+
+    public String getHostname() {
+        return hostname;
+    }
+
+    public void setHostname(String hostname) {
+        this.hostname = hostname;
+    }
+
+    public String getStartScript() {
+        return startScript;
+    }
+
+    public void setStartScript(String startScript) {
+        this.startScript = startScript;
+    }
+
+    public Map<String, DependencyComponent> getDependencies() {
+        return dependencies;
+    }
+
+    public void setDependencies(Map<String, DependencyComponent> dependencies) {
+        this.dependencies = dependencies;
+    }
+
+    public List<String> getEnvironment() {
+        return environment;
+    }
+
+    public void setEnvironment(List<String> environment) {
+        this.environment = environment;
+    }
+
+    public List<String> getExpose() {
+        return expose;
+    }
+
+    public void setExpose(List<String> expose) {
+        this.expose = expose;
+    }
+
+    public List<String> getDepends_on() {
+        return depends_on;
+    }
+
+    public void setDepends_on(List<String> depends_on) {
+        this.depends_on = depends_on;
+    }
+}
