+++ test/plugin/runner-helper/src/main/resources/container-start-script.template
