+++ test/plugin/runner-helper/src/main/resources/scenario.sh
