+++ test/plugin/runner-helper/src/test/resources/configuration-test.yml
@@ -1,0 +1,31 @@
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+type: jvm
+entryService: http://localhost:8080/solrj-7.x-scenario/case/solrj
+healthCheck: http://localhost:8080/solrj-7.x-scenario/heathcheck
+startScript: ./bin/startup.sh
+framework: solrj
+environment:
+  - SOLR_SERVER=http://solr-server:8983
+dependencies:
+  solr:
+    image: solr:${CASE_SERVER_IMAGE_VERSION}
+    hostname: solr-server
+    entrypoint:
+      - docker-entrypoint.sh
+      - solr-precreate
+      - mycore
+
