--- test/plugin/runner/src/test/resources/container-start-script.template
@@ -1,29 +1,0 @@
-<#--
-  ~ Licensed to the Apache Software Foundation (ASF) under one or more
-  ~ contributor license agreements.  See the NOTICE file distributed with
-  ~ this work for additional information regarding copyright ownership.
-  ~ The ASF licenses this file to You under the Apache License, Version 2.0
-  ~ (the "License"); you may not use this file except in compliance with
-  ~ the License.  You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
--->
-
-docker run -itd  \
-        --name skywalking-agent-test-${scenario_name}-${scenario_version}  \
-        --env SCENARIO_VERSION=${scenario_version}  \
-        --env SCENARIO_SUPPORT_FRAMEWORK=${scenario_name} \
-        --env SCENARIO_ENTRY_SERVICE=${entry_service}  \
-        --env SCENARIO_HEALTH_CHECK_URL=${health_check} \
-        -v ${agent_home}:/usr/local/skywalking-agent-scenario/agent \
-        -v ${scenario_home}/logs:/usr/local/skywalking-agent-scenario/logs \
-        -v ${scenario_home}/reports:/usr/local/skywalking-agent-scenario/reports \
-        -v ${scenario_home}/data:/usr/local/skywalking-agent-scenario/data \
-        -v ${scenario_home}/packages:/usr/local/skywalking-agent-scenario/packages \
-        ${docker_image_name}:latest
