--- test/plugin/runner/src/test/resources/docker-compose.template
@@ -1,58 +1,0 @@
-<#--
-  ~ Licensed to the Apache Software Foundation (ASF) under one or more
-  ~ contributor license agreements.  See the NOTICE file distributed with
-  ~ this work for additional information regarding copyright ownership.
-  ~ The ASF licenses this file to You under the Apache License, Version 2.0
-  ~ (the "License"); you may not use this file except in compliance with
-  ~ the License.  You may obtain a copy of the License at
-  ~
-  ~     http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
--->
-version: 2.1
-services:
-    ${docker_container_name}:
-        image: ${docker_image_name}:latest
-        <#if links?size \gt 0>
-        links:
-        <#list links as link>
-            - ${link}
-        </#list>
-        </#if>
-        environment:
-            - SERVER_ADDR:${server_addr}
-            - SCENARIO_NAME:${scenario_name}
-            - SCENARIO_VERSION:${scenario_version}
-            - SCENARIO_ENTRY_SERVICE:${entry_service}
-            - SCENARIO_HEALTH_CHECK_URL:${health_check}
-        volumes:
-            - ${agent_home}:/usr/local/skywalking-agent-scenario/agent
-            - ${scenario_home}:/usr/local/skywalking-agent-scenario
-<#list services as service>
-    ${service.name}:
-        image: ${service.image}:${service.version}
-        host: ${service.host}
-        <#if service.volumes?size \gt 0>
-        volumes:
-        <#list service.volumes as volume>
-            - ${volume}
-        </#list>
-        </#if>
-        <#if service.environments?size \gt 0>
-        environments:
-        <#list service.environments as environment>
-            - ${environment}
-        </#list>
-        </#if>
-        <#if service.expose?size \gt 0>
-        expose:
-        <#list service.expose as expose>
-            - ${expose}
-        </#list>
-        </#if>
-</#list>
