+++ test/plugin/scenarios/solrj-7.x-scenario/config/expectedData.yaml
@@ -1,0 +1,181 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+registryItems:
+  applications:
+    - {solrj-7.x-scenario: nq 0}
+  instances:
+    - {solrj-7.x-scenario: nq 0}
+  operationNames:
+    - solrj-7.x-scenario: [/solrj-scenario/case/solrj, solrJ/mycore/update/COMMIT, solrJ/mycore/update/DELETE_BY_IDS,
+                           solrJ/mycore/select, solrJ/mycore/update/DELETE_BY_QUERY, solrJ/mycore/get,
+                           /solrj-scenario/case/healthcheck, solrJ/mycore/update/OPTIMIZE, solrJ/mycore/update/ADD]
+  heartbeat: []
+segmentItems:
+  - applicationCode: solrj-7.x-scenario
+    segmentSize: 2
+    segments:
+      - segmentId: not null
+        spans:
+          - operationName: /solrj-scenario/case/healthcheck
+            operationId: 0
+            parentSpanId: -1
+            spanId: 0
+            spanLayer: Http
+            startTime: gt 0
+            endTime: gt 0
+            componentId: 1
+            componentName: ''
+            isError: false
+            spanType: Entry
+            peer: ''
+            peerId: 0
+            tags:
+              - {key: url, value: 'http://localhost:8080/solrj-scenario/case/healthcheck'}
+              - {key: http.method, value: HEAD}
+      - segmentId: not null
+        spans:
+          - operationName: solrJ/mycore/update/ADD
+            operationId: 0
+            parentSpanId: 0
+            spanId: 1
+            spanLayer: Database
+            startTime: gt 0
+            endTime: gt 0
+            componentId: 63
+            componentName: ''
+            isError: false
+            spanType: Exit
+            peer: solr-server:8983
+            peerId: 0
+            tags:
+              - {key: db.type, value: Solr}
+              - {key: QTime, value: gt 0}
+          - operationName: solrJ/mycore/update/COMMIT
+            operationId: 0
+            parentSpanId: 0
+            spanId: 2
+            spanLayer: Database
+            startTime: gt 0
+            endTime: gt 0
+            componentId: 63
+            componentName: ''
+            isError: false
+            spanType: Exit
+            peer: solr-server:8983
+            peerId: 0
+            tags:
+              - {key: db.type, value: Solr}
+              - {key: QTime, value: gt 0}
+          - operationName: solrJ/mycore/update/OPTIMIZE
+            operationId: 0
+            parentSpanId: 0
+            spanId: 3
+            spanLayer: Database
+            startTime: gt 0
+            endTime: gt 0
+            componentId: 63
+            componentName: ''
+            isError: false
+            spanType: Exit
+            peer: solr-server:8983
+            peerId: 0
+            tags:
+              - {key: db.type, value: Solr}
+              - {key: QTime, value: gt 0}
+          - operationName: solrJ/mycore/select
+            operationId: 0
+            parentSpanId: 0
+            spanId: 4
+            spanLayer: Database
+            startTime: gt 0
+            endTime: gt 0
+            componentId: 63
+            componentName: ''
+            isError: false
+            spanType: Exit
+            peer: solr-server:8983
+            peerId: 0
+            tags:
+              - {key: db.type, value: Solr}
+              - {key: start, value: '0'}
+              - {key: qt, value: /select}
+              - {key: numFound, value: '100'}
+          - operationName: solrJ/mycore/get
+            operationId: 0
+            parentSpanId: 0
+            spanId: 5
+            spanLayer: Database
+            startTime: gt 0
+            endTime: gt 0
+            componentId: 63
+            componentName: ''
+            isError: false
+            spanType: Exit
+            peer: solr-server:8983
+            peerId: 0
+            tags:
+              - {key: db.type, value: Solr}
+              - {key: start, value: '0'}
+              - {key: qt, value: /get}
+              - {key: numFound, value: '1'}
+          - operationName: solrJ/mycore/update/DELETE_BY_IDS
+            operationId: 0
+            parentSpanId: 0
+            spanId: 6
+            spanLayer: Database
+            startTime: gt 0
+            endTime: gt 0
+            componentId: 63
+            componentName: ''
+            isError: false
+            spanType: Exit
+            peer: solr-server:8983
+            peerId: 0
+            tags:
+              - {key: db.type, value: Solr}
+              - {key: QTime, value: 'gt 0'}
+          - operationName: solrJ/mycore/update/DELETE_BY_QUERY
+            operationId: 0
+            parentSpanId: 0
+            spanId: 7
+            spanLayer: Database
+            startTime: gt 0
+            endTime: gt 0
+            componentId: 63
+            componentName: ''
+            isError: false
+            spanType: Exit
+            peer: solr-server:8983
+            peerId: 0
+            tags:
+              - {key: db.type, value: Solr}
+              - {key: QTime, value: 'gt 0'}
+          - operationName: /solrj-scenario/case/solrj
+            operationId: 0
+            parentSpanId: -1
+            spanId: 0
+            spanLayer: Http
+            startTime: gt 0
+            endTime: gt 0
+            componentId: 1
+            componentName: ''
+            isError: false
+            spanType: Entry
+            peer: ''
+            peerId: 0
+            tags:
+              - {key: url, value: 'http://localhost:8080/solrj-scenario/case/solrj'}
+              - {key: http.method, value: GET}
